Start Of ColorSystem Implementation.
A Color system that will control how color interaction in-game will work.

ColorHolder will hold the proper material (will be changed to a Material Property Block) and eColor type.

eColor is the enum for in-game color types.

ColorReceiver is the class that handle the Receiving of colors which then (if applicable) will call IColorReactor which will cause a proper reaction.

BeehiveColorReactor and NectarColorReactor both implement IColorReactor.