Added Knockback and Polished ColorSpreader system
Implemented (naively)  player knockback -
Knockbacked method in player controller which is called and starts and enum for rotation with force.
This is a bad way of implementation and will probably be replace with a State Design Pattern if I will implement more "Player States"

ColorSpreader gives the ability to have another group of objects for a secondary color spread.